DivCounterManager/requirements.txt
grpcio>=1.64.1
grpcio-tools>=1.64.1
uvicorn>=0.30.1
fastapi>=0.111.0
